Cast (height - 1) to ptrdiff_t in vpx_img_flip()
vpx_img_flip() casts (height - 1) to `signed` before multiplying it with
stride. Change the `signed` casts to `ptrdiff_t` casts to avoid
overflows on 64-bit systems.

Since the img->planes[] buffers have been allocated successfully, the
(height - 1) * stride products can be represented in the `ptrdiff_t`
type, so the `ptrdiff_t` casts won't truncate (height - 1).

+++ vpx/src/vpx_image.c
@@ -10,6 +10,7 @@
 
 #include <assert.h>
 #include <limits.h>
+#include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
 
@@ -249,26 +250,21 @@ void vpx_img_flip(vpx_image_t *img) {
   const unsigned int chroma_height =
       (img->d_h + img->y_chroma_shift) >> img->y_chroma_shift;
 
-  /* Note: In the calculation pointer adjustment calculation, we want the
-   * rhs to be promoted to a signed type. Section 6.3.1.8 of the ISO C99
-   * standard indicates that if the adjustment parameter is unsigned, the
-   * stride parameter will be promoted to unsigned, causing errors when
-   * the lhs is a larger type than the rhs.
-   */
-  img->planes[VPX_PLANE_Y] += (signed)(img->d_h - 1) * img->stride[VPX_PLANE_Y];
+  img->planes[VPX_PLANE_Y] +=
+      (ptrdiff_t)(img->d_h - 1) * img->stride[VPX_PLANE_Y];
   img->stride[VPX_PLANE_Y] = -img->stride[VPX_PLANE_Y];
 
   img->planes[VPX_PLANE_U] +=
-      (signed)(chroma_height - 1) * img->stride[VPX_PLANE_U];
+      (ptrdiff_t)(chroma_height - 1) * img->stride[VPX_PLANE_U];
   img->stride[VPX_PLANE_U] = -img->stride[VPX_PLANE_U];
 
   img->planes[VPX_PLANE_V] +=
-      (signed)(chroma_height - 1) * img->stride[VPX_PLANE_V];
+      (ptrdiff_t)(chroma_height - 1) * img->stride[VPX_PLANE_V];
   img->stride[VPX_PLANE_V] = -img->stride[VPX_PLANE_V];
 
   if (img->fmt & VPX_IMG_FMT_HAS_ALPHA) {
     img->planes[VPX_PLANE_ALPHA] +=
-        (signed)(img->d_h - 1) * img->stride[VPX_PLANE_ALPHA];
+        (ptrdiff_t)(img->d_h - 1) * img->stride[VPX_PLANE_ALPHA];
     img->stride[VPX_PLANE_ALPHA] = -img->stride[VPX_PLANE_ALPHA];
   }
 }
